Moved the newly components into folders to match existing folder structure. Replaced div with a react fragment. Beautified the menu in Navbar.
