+++ client/src/pages/Dashboard/Dashboard.tsx
@@ -3,8 +3,8 @@ import { useAuth } from '../../context/useAuthContext';
 import { useSocket } from '../../context/useSocketContext';
 import { useHistory } from 'react-router-dom';
 import { CircularProgress, Typography, Grid } from '@mui/material';
-import { Navbar } from '../../components/Navbar';
-import { RandomFactCard } from '../../components/RandomFactCard';
+import { Navbar } from '../../components/Navbar/Navbar';
+import { RandomFactCard } from '../../components/FactCard/RandomFactCard';
 import { facts } from '../../mocks/mockFacts';
 
 export default function Dashboard(): JSX.Element {
